+++ client-app/src/desktop/tabs/containers/tab/tabs/DynamicTabContainerPanel.js
