+++ client-app/src/desktop/tabs/containers/tab/tabs/index.js
@@ -1,0 +1,3 @@
+export * from './TabStateContainerPanel';
+export * from './DynamicTabContainerPanel';
+export * from './CustomTabContainerPanel';
